+++ src/oersted/magnetization.py
@@ -1,5 +1,7 @@
 """Operations for magnetic materials"""
 
+from oersted import DirectSolver, OctreeSolver
+
 from numpy.typing import NDArray
 from numpy import float64
 
@@ -9,14 +11,7 @@ from ._oersted import magnetization_tet4
 
 
 def demag_tet4(
-    mesh: Mesh,
-    material: Material,
-    h_external: NDArray[float64],
-    max_iterations: int = 50,
-    tol: float = 1.0,
-    theta: float = 0.5,
-    leaf_threshold: int = 0,
-    nthreads_requested: int = 0,
+    mesh: Mesh, material: Material, h_external: NDArray[float64], solver: DirectSolver | OctreeSolver
 ) -> tuple[NDArray[float64], NDArray[float64]]:
     """Compute magnetization field M and the total H field at element centroids, given a background field
 
@@ -28,13 +23,32 @@ def demag_tet4(
             these are indices of the array `nodes`, not of the solver's node numbers
         material: linear or nonlinear magnetic maaterial properties
         h_external: (Ne,3) external field at each element centroid
-        max_iterations: number of solver iterations before exit
-        tol: maximum amount of change per individual component of M at each element
+        solver: solution parameters for the problem, including iteration method
 
     Returns:
         (M, Htotal): each (Ne, 3), magnetization field M(Htotal) and total H field at element
-            centroids. These can be summed to give B = mu0 * (Htotal + M)
+            centroids. These can be summed to give B = mu0 * (Htotal + M).
     """
+
+    theta: float
+    leaf_threshold: int
+
+    if isinstance(solver, DirectSolver):
+        theta = 0.5
+        leaf_threshold = 0
+    else:
+        theta = solver.theta
+        leaf_threshold = solver.leaf_threshold
+
     return magnetization_tet4(
-        mesh.nodes, mesh.connectivity, material.chi(1.0), h_external, tol, max_iterations, theta, leaf_threshold, nthreads_requested
+        mesh.nodes,
+        mesh.connectivity,
+        material.chi(1.0),
+        h_external,
+        solver.tol,
+        solver.max_iterations,
+        theta,
+        leaf_threshold,
+        solver.alpha,
+        solver.n_threads,
     )